feat(ui): client-side FDF-driven in-game HUD (ConsoleUI screen controller)
Replace server-authored svc_layout HUD with client-side FDF rendering.
Loads Blizzard's ConsoleUI/ResourceBar/UpperButtonBar/InfoPanel FDF
files from MPQ at runtime and draws through UI_DrawFrames().

Key changes:
- stb_fdf.h: shared declarations header for FDF types/APIs
- screens/console_ui.c: screen controller with load/init/refresh/draw
- Generated binding headers for 6 ConsoleUI FDF files
- UI_EnterGameMode() wires ConsoleUI into game mode
- Stubbed server HUD functions (Get_Commands_f, Get_Portrait_f,
  G_UpdateClientInfoPanels, G_UpdateClientResourceBars, etc.)
- Updated AGENTS.md and ui-authoring.md with new patterns

Server still sends svc_layout for FDF-based skill/build menus
(UI_WRITE_LAYER) and data the client cannot compute (minimap,
portrait model). The old UIFRAME render path in cl_unit_layout.c
is preserved for those cases.
