+++ games/warcraft-3/ui/ui_local.h
@@ -5,334 +5,23 @@
  * constants used within the UI library. External code should only include
  * ui.h, never this file.
  *
- * Frame template structures (FRAMEDEF, frameDef_t) are defined here.
- * These are used internally by the UI library to build frame hierarchies
- * from FDF files.
+ * Frame template structures (FRAMEDEF) are defined in stb_fdf.h which is
+ * shared with the game module. This header adds UI-specific extensions.
  */
 #ifndef ui_local_h
 #define ui_local_h
 
 #include <stdio.h>
 
+#include "stb_fdf.h"
 #include "client/ui.h"
-#include "common/shared.h"  /* For PLAYERSTATE/PLAYERTEXT enums */
 #include "games/warcraft-3/common/mapinfo.h"
 
-/* Constants */
-#define MAX_BUILD_QUEUE 7
-/* MAX_SELECTED_ENTITIES is defined in common/shared.h */
-#ifndef MAX_UI_CLASSES
-#define MAX_UI_CLASSES 4096
-#endif
-
-#define UI_BASE_WIDTH 0.8f
-#define UI_BASE_HEIGHT 0.6f
-#define UI_MIN_ASPECT (4.0f / 3.0f)
-
 /* Forward declarations */
-typedef struct uiFrameDef_s frameDef_t;
-typedef frameDef_t FRAMEDEF;  /* Also available as non-pointer type */
-typedef frameDef_t *LPFRAMEDEF;
-typedef frameDef_t const *LPCFRAMEDEF;
 typedef struct uiScreen_s uiScreen_t;  /* Defined in ui_screen.h */
 
 /* Global import callbacks (filled by UI_GetAPI) */
 extern uiImport_t uiimport;
-
-/* Frame point positioning */
-typedef enum {
-    FRAMEPOINT_TOPLEFT,
-    FRAMEPOINT_TOP,
-    FRAMEPOINT_TOPRIGHT,
-    FRAMEPOINT_UNUSED1,
-    FRAMEPOINT_LEFT,
-    FRAMEPOINT_CENTER,
-    FRAMEPOINT_RIGHT,
-    FRAMEPOINT_UNUSED2,
-    FRAMEPOINT_BOTTOMLEFT,
-    FRAMEPOINT_BOTTOM,
-    FRAMEPOINT_BOTTOMRIGHT,
-    FRAMEPOINT_UNUSED3,
-} UIFRAMEPOINT;
-
-typedef enum {
-    FONTFLAGS_FIXEDSIZE,
-    FONTFLAGS_PASSWORDFIELD,
-} UIFONTFLAGS;
-
-typedef enum {
-    FILETEXTURE,
-} HIGHLIGHTTYPE;
-
-typedef enum {
-    AUTOTRACK = 1,
-    HIGHLIGHTONFOCUS = 2,
-    HIGHLIGHTONMOUSEOVER = 4,
-} CONTROLSTYLE;
-
-typedef enum {
-    LAYOUT_HORIZONTAL,
-    LAYOUT_VERTICAL,
-} LAYOUTDIRECTION;
-
-typedef struct {
-    UINAME frame;
-    UINAME text;
-} BUTTONTEXT;
-
-typedef struct {
-    uiFramePointPos_t targetPos;
-    bool used;
-    LPCFRAMEDEF relativeTo;
-    FLOAT offset;
-} FRAMEPOINT;
-
-typedef FRAMEPOINT const *LPCFRAMEPOINT;
-
-#define UI_MAX_MAP_LIST_ITEMS 1024
-#define UI_MAX_MENU_ITEMS 32
-
-typedef struct {
-    UINAME text;
-    LONG value;
-} uiMenuItem_t;
-
-typedef struct {
-    PATHSTR path;
-    char name[128];
-    char description[512];
-    char suggestedPlayers[96];
-    char mapSize[32];
-    char tileset[64];
-    DWORD players;
-    DWORD flags;
-} uiMapListItem_t;
-
-typedef struct {
-    uiMapListItem_t items[UI_MAX_MAP_LIST_ITEMS];
-    DWORD count;
-    DWORD selected;
-    DWORD scroll;
-    FLOAT visualScroll;
-} uiMapListState_t;
-
-typedef struct {
-    uiMapListState_t *State;
-    DWORD VisibleRows;
-    FLOAT RowHeight;
-    FLOAT InsetX;
-    FLOAT InsetY;
-    UINAME SelectCommand;
-    UINAME FontName;
-    FLOAT FontSize;
-    COLOR32 TextColor;
-    COLOR32 SelectedTextColor;
-} uiMapListControl_t;
-
-/* UI interaction flags for uiFrameDef_s.ui_flags */
-#define UIFLAG_PRESSED  (1 << 0)
-#define UIFLAG_HOVERED  (1 << 1)
-#define UIFLAG_CHECKED  (1 << 2)
-#define UIFLAG_DISABLED (1 << 3)
-#define UIFLAG_ACTIVE   (1 << 4)
-#define UIFLAG_VISIBLE  (1 << 5)
-#define UIFLAG_PASSTHROUGH (1 << 6)  /* child of popup: rendered but not hit-tested */
-
-/* Frame template definition (server-side/library-side only) */
-struct uiFrameDef_s {
-    LPCFRAMEDEF Parent;
-    FRAMETYPE Type;
-    UINAME Name;
-    UINAME TextStorage;
-    UINAME OnClick;
-    LPCSTR Text, Tip, Ubertip;
-    FLOAT Width, Height;
-    COLOR32 Color;
-    BLEND_MODE AlphaMode;
-    BOOL DecorateFileNames;
-    BOOL inuse;
-    BOOL AnyPointsSet;
-    BOOL hidden;
-    BOOL disabled;
-    DWORD TextLength;
-    DWORD Stat;
-    LPSTR DynamicText;
-    DWORD DynamicTextCapacity;
-    struct {
-        FRAMEPOINT x[FPP_COUNT];
-        FRAMEPOINT y[FPP_COUNT];
-    } Points;
-    struct {
-        DWORD Image;
-        DWORD Image2;
-        BOX2 TexCoord;
-    } Texture;
-    struct {
-        BOOL TileBackground;
-        DWORD Background;
-        DWORD CornerFlags;
-        FLOAT CornerSize;
-        FLOAT BackgroundSize;
-        FLOAT BackgroundInsets[4];
-        DWORD EdgeFile;
-        BOOL BlendAll;
-        BOOL Mirrored;
-    } Backdrop;
-    UINAME DialogBackdropName;
-    LPCFRAMEDEF DialogBackdrop;
-    struct {
-        DWORD model;
-    } Portrait;
-    struct {
-        UIFRAMEPOINT corner;
-        FLOAT x, y;
-    } Anchor;
-    struct {
-        UIFRAMEPOINT type;
-        LPCFRAMEDEF relativeTo;
-        UIFRAMEPOINT target;
-        FLOAT x, y;
-    } SetPoint;
-    struct {
-        UINAME Name;
-        UINAME Unknown;
-        UIFONTFLAGS FontFlags;
-        FLOAT Size;
-        DWORD Index;
-        COLOR32 Color;
-        COLOR32 HighlightColor;
-        COLOR32 DisabledColor;
-        COLOR32 ShadowColor;
-        VECTOR2 ShadowOffset;
-        struct {
-            VECTOR2 Offset;
-            uiFontJustificationH_t Horizontal;
-            uiFontJustificationV_t Vertical;
-        } Justification;
-    } Font;
-    struct {
-        HIGHLIGHTTYPE Type;
-        DWORD AlphaFile;
-        BLEND_MODE AlphaMode;
-        COLOR32 Color;
-    } Highlight;
-    struct {
-        VECTOR2 PushedTextOffset;
-        UINAME NormalTexture;
-        UINAME PushedTexture;
-        UINAME DisabledTexture;
-        UINAME UseHighlight;
-        BUTTONTEXT NormalText;
-        BUTTONTEXT DisabledText;
-        BUTTONTEXT HighlightText;
-    } Button;
-    struct {
-        DWORD Style;
-        struct {
-            UINAME Normal;
-            UINAME Pushed;
-            UINAME Disabled;
-            UINAME MouseOver;
-            UINAME DisabledPushed;
-            UINAME Focus;
-        } Backdrop;
-        UINAME ShortcutKey;
-        UINAME TabFocusNext;
-        BOOL TabFocusDefault;
-    } Control;
-    struct {
-        FLOAT InitialValue;
-        LAYOUTDIRECTION Layout;
-        FLOAT MaxValue;
-        FLOAT MinValue;
-        FLOAT StepSize;
-        UINAME ThumbButtonFrame;
-        UINAME IncButtonFrame;
-        UINAME DecButtonFrame;
-    } Slider;
-    struct {
-        FLOAT Border;
-        UINAME ScrollBar;
-        UINAME FetchCommand;
-    } ListBox;
-    uiMapListControl_t MapListControl;
-    struct {
-        FLOAT Border;
-        struct {
-            UINAME Text;
-            DWORD Value;
-            FLOAT Height;
-        } Item;
-        DWORD ItemCount;
-        uiMenuItem_t Items[UI_MAX_MENU_ITEMS];
-        COLOR32 TextHighlightColor;
-    } Menu;
-    struct {
-        FLOAT BorderSize;
-        COLOR32 CursorColor;
-        COLOR32 HighlightColor;
-        BOOL HighlightInitial;
-        DWORD MaxChars;
-        BOOL Focus;
-        UINAME Text;
-        COLOR32 TextColor;
-        UINAME TextFrame;
-        VECTOR2 TextOffset;
-    } Edit;
-    struct {
-        UINAME ArrowFrame;
-        UINAME MenuFrame;
-        UINAME TitleFrame;
-        FLOAT ButtonInset;
-    } Popup;
-    struct {
-        FLOAT LineHeight;
-        FLOAT LineGap;
-        FLOAT Inset;
-        DWORD MaxLines;
-        UINAME ScrollBar;
-    } TextArea;
-    struct {
-        UINAME CheckHighlight;
-        UINAME DisabledCheckHighlight;
-        BOOL Checked;
-    } CheckBox;
-    struct {
-        LPCFRAMEDEF FirstItem;
-        LPCFRAMEDEF BuildTimer;
-        FLOAT ItemOffset;
-        DWORD NumQueue;
-        uiBuildQueueItem_t Queue[MAX_BUILD_QUEUE];
-    } BuildQueue;
-    struct {
-        DWORD HpBar;
-        DWORD ManaBar;
-        VECTOR2 Offset;
-        DWORD NumColumns;
-        DWORD NumItems;
-        uiMultiselectItem_t Items[MAX_SELECTED_ENTITIES];
-    } Multiselect;
-    /* Interaction state — updated by event handler, read by draw */
-    DWORD ui_flags;
-    /* Per-type event handler: called from UI_MouseEventLocal */
-    void (*event_handler)(LPFRAMEDEF frame, uiMouseEvent_t event, FLOAT fdf_x, FLOAT fdf_y, int32_t param);
-    /* Per-type draw function: called from UI_DrawFrameOne */
-    void (*draw)(LPCFRAMEDEF frame, LPCRECT rect);
-};
-
-/* Global parsed FDF frame table. */
-extern FRAMEDEF frames[MAX_UI_CLASSES];
-
-/* Stub types for server-side code compatibility (not used in client UI library) */
-typedef struct {} uiTrigger_t;
-typedef void *LPEDICT;
-
-/* Convenience macros for frame lookup */
-#define UI_FRAME_GLOBAL(NAME) LPFRAMEDEF NAME = UI_FindFrame(#NAME);
-#define UI_FRAME_CHILD(PARENT, NAME) LPFRAMEDEF NAME = UI_FindChildFrame(PARENT, #NAME);
-#define UI_FRAME_SELECT(_1, _2, NAME, ...) NAME
-#define UI_FRAME(...) UI_FRAME_SELECT(__VA_ARGS__, UI_FRAME_CHILD, UI_FRAME_GLOBAL)(__VA_ARGS__)
-#define UI_CHILD_FRAME(NAME, PARENT) LPFRAMEDEF NAME = UI_FindChildFrame(PARENT, #NAME);
 
 /* Internal function prototypes */
